formatDateをcommonFunctionに移行
