+++ app/components/ArticleCard.tsx
@@ -1,4 +1,5 @@
 import { Link } from "@remix-run/react";
+import { formatDate } from "../utils/commonFunction";
 
 type Article = {
   id: number;
@@ -11,16 +12,6 @@ type Article = {
   author: { profile: { name: string } | null };
   favoritedBy: { id: number; userId: number; articleId: number }[];
 };
-
-function formatDate(date: Date): string {
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, "0"); // 月は0から始まるため+1
-  const day = String(date.getDate()).padStart(2, "0");
-  const hours = String(date.getHours()).padStart(2, "0");
-  const minutes = String(date.getMinutes()).padStart(2, "0");
-
-  return `${year}/${month}/${day} ${hours}:${minutes}`;
-}
 
 export default function ArticleCard({ article }: { article: Article }) {
   return (
