+++ app/utils/commonFunction.ts
@@ -1,0 +1,14 @@
+/**
+ * Date型を yyyy/mm/dd hh:mm に変換
+ * @param date
+ * @returns
+ */
+export function formatDate(date: Date): string {
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, "0"); // 月は0から始まるため+1
+  const day = String(date.getDate()).padStart(2, "0");
+  const hours = String(date.getHours()).padStart(2, "0");
+  const minutes = String(date.getMinutes()).padStart(2, "0");
+
+  return `${year}/${month}/${day} ${hours}:${minutes}`;
+}
